Show assigned roles in a principal table
